Combine two views into one
