+++ bonsai/Views/FlexibleView/FlexibleView.swift
@@ -3,14 +3,14 @@
 
 import SwiftUI
 
-/// Facade of our view, its main responsibility is to get the available width
-/// and pass it down to the real implementation, `_FlexibleView`.
 struct FlexibleView<Data: Collection, Content: View>: View where Data.Element: Hashable {
    let data: Data
    let spacing: CGFloat
    let alignment: HorizontalAlignment
    let content: (Data.Element) -> Content
    @State private var availableWidth: CGFloat = 0
+   @State var elementsSize: [Data.Element: CGSize] = [:]
+
    
    var body: some View {
       ZStack(alignment: Alignment(horizontal: alignment, vertical: .center)) {
@@ -20,13 +20,42 @@ struct FlexibleView<Data: Collection, Content: View>: View where Data.Element: H
                availableWidth = size.width
             }
          
-         _FlexibleView(
-            availableWidth: availableWidth,
-            data: data,
-            spacing: spacing,
-            alignment: alignment,
-            content: content
-         )
+          
+          VStack(alignment: alignment, spacing: spacing) {
+             ForEach(computeRows(), id: \.self) { rowElements in
+                HStack(spacing: spacing) {
+                   ForEach(rowElements, id: \.self) { element in
+                      content(element)
+                         .fixedSize()
+                         .readSize { size in
+                            elementsSize[element] = size
+                         }
+                   }
+                }
+             }
+          }
       }
    }
+    
+    func computeRows() -> [[Data.Element]] {
+       var rows: [[Data.Element]] = [[]]
+       var currentRow = 0
+       var remainingWidth = availableWidth
+
+       for element in data {
+          let elementSize = elementsSize[element, default: CGSize(width: availableWidth, height: 1)]
+
+          if remainingWidth - (elementSize.width + spacing) >= 0 {
+             rows[currentRow].append(element)
+          } else {
+             currentRow = currentRow + 1
+             rows.append([element])
+             remainingWidth = availableWidth
+          }
+
+          remainingWidth = remainingWidth - (elementSize.width + spacing)
+       }
+
+       return rows
+    }
 }
